requirements.txt
tkinter
psutil>=5.9.0
matplotlib>=3.7.0
gputil>=1.4.0
tk>=0.1.0
pandas>=2.0.0
pillow>=10.0.0
pystray>=0.19.0
